% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_funs.R
\name{access_types}
\alias{access_types}
\alias{collections}
\alias{country_codes}
\title{Small Helper Functions}
\usage{
access_types(org = c("wb", "fao", "unhcr", "ihsn", "ilo"))

collections(org = c("wb", "fao", "unhcr", "ihsn", "ilo"))

country_codes(org = c("wb", "fao", "unhcr", "ihsn", "ilo"))
}
\arguments{
\item{org}{a character string (default = "wb") specifying the name of an organization.}
}
\value{
A data frame
}
\description{
Small helper functions to get available data access codes, collections, and
country codes from different organization repositories.
}
\examples{

access_types("fao")
collections("ilo")
country_codes("wb")

}
\author{
Gutama Girja Urago
}
