% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq_test}
\alias{chisq_test}
\title{Weighted Chi-Square Test}
\usage{
chisq_test(x, y, w = NULL, data = NULL)
}
\arguments{
\item{x}{A categorical variable to be used as rows in the contingency table.}

\item{y}{A categorical variable to be used as columns in the contingency table.}

\item{w}{An optional weighting variable.}

\item{data}{A data frame or matrix containing the variables.}
}
\value{
A list with the results of the chi-square test, including:
\describe{
\item{\code{statistic}}{The value of the chi-square statistic.}
\item{\code{parameter}}{The degrees of freedom for the test.}
\item{\code{p.value}}{The p-value of the test.}
\item{\code{n.cases}}{The number of cases used in the test.}
}
}
\description{
Performs a weighted chi-square test on two categorical variables, with an optional weighting variable.
}
\details{
This function calculates the chi-square statistic for two categorical variables, taking into account optional weights.
}
\examples{
chisq_test(x = gender, y = education, w = w, data = micro_data)
}
