% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Get Collections}
\usage{
collections(org = "wb")
}
\arguments{
\item{org}{a character string (default = "wb") specifying the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}
}
\value{
A data frame that contains all available collections
}
\description{
Get a list of all the collections the organization offers.
}
\examples{
collections("fao")
}
\author{
Gutama Girja Urago
}
