% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mean.R
\name{compare_mean}
\alias{compare_mean}
\title{Compare Group (Weighted) Mean}
\usage{
compare_mean(
  data,
  vars,
  by,
  w = NULL,
  digits = 2,
  use_labels = TRUE,
  stars_on = "stat",
  sign_levels = c(`***` = 0.01, `**` = 0.05, `*` = 0.1)
)
}
\arguments{
\item{data}{is a data frame that contains all variables, including weight, if supplied.}

\item{vars}{is a vector of variables for which means are supposed to be calculated.}

\item{by}{is a factor grouping variable. It internally converts to a factor
if another type of variable is supplied.}

\item{w}{is an optional weight variable.}

\item{digits}{is an integer indicating number of digits after decimal point.}

\item{use_labels}{is a logical indicating if the variable labels should be
returned instead of variable names.}

\item{stars_on}{is a character indicating where to paste significance stars
taking value of 'non', 'stat' or 'diffs'. If 'none' p values are returned as a
separate column.}

\item{sign_levels}{is a named vector of p_values with corresponding stars.}
}
\value{
Returns a data frame ready to be exported or manipulated with fancy
gt table. The 'diffs' column is returned only when a grouping variable has only
two levels.
}
\description{
Create a summary table for the (weighted) mean by group and performs an ANOVA test.
If the grouping variable is binary, t-statistics is reported instead of f-statistic.
}
\details{
This function calculates the mean using \code{collapse::fmean()} by using a grouping
variable. After all mean values are calculated, it employs the \code{stats::lm()} model
to obtain test statistics. If the grouping variable is a factor with two levels,
the square root of the resulting F statistics is returned. Therefore, for binary
grouping variables, the test is a two-sample t-test with an assumption of equal variance.
}
\examples{

if (FALSE) {
compare_mean(iris, vars = c(Sepal.Length, Sepal.Width,
Petal.Length, Petal.Width),
by = Species)
}

}
\seealso{
\code{\link[=dummify]{dummify()}} to create dummy variables
}
\author{
Gutama Girja Urago
}
