% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{country_codes}
\alias{country_codes}
\title{Country Codes}
\usage{
country_codes(org = "wb")
}
\arguments{
\item{org}{a string that represents the name of an organization.}
}
\value{
A data frame with all country codes
}
\description{
Get a list of all country codes
}
\examples{
country_codes()
}
\author{
Gutama Girja Urago
}
