% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{data_files}
\alias{data_files}
\title{Study Data Files}
\usage{
data_files(id, org = c("wb", "fao", "unhcr", "ihsn", "ilo"))
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.
See object returned by \code{search_catalog()} or \code{latest_entries()}}

\item{org}{a string that represents the name of an organization.}
}
\value{
A data frame
}
\description{
Get list of all data files included in the study
}
\examples{
if (FALSE){
data_files("ALB_2012_LSMS_v01_M_v01_A_PUF", "wb")
}
}
\author{
Gutama Girja Urago
}
