% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Extracts web data dictionary}
\usage{
dictionary(catalog, filename, org = "wb", namecol = TRUE, format = TRUE)
}
\arguments{
\item{catalog}{integer in the web link after catalog without "/"}

\item{filename}{string in the web link after data-dictionary without "/". It
usually starts with capital letter followed by integer. But it is better to
include full string after the capital letter, especially when extracting dictionary of
many datasets since the last string helps to identify the corresponding dataset.}

\item{org}{name of an organization "wb", "fao", "unhcr" or "ihsn", \code{default = "wb"}}

\item{namecol}{if \code{TRUE} empty column for new name is mutated.
\code{default = TRUE}, then manually set new names in excel sheet.}

\item{format}{Cleans preceding integers in variable descriptions}
}
\value{
a data table
}
\description{
Extracts data dictionary from World bank, FAO, UNHCR and IHSN microdata catalogs given
catalog number and file name.
}
\details{
If there is any missing description of variables, it automatically
set "Unknown" to the description to write balanced data table
}
\examples{
dictionary(3823, "F2")
dictionary(catalog = 5984, filename = "F5")
dictionary(catalog = 1771, filename = "F392", org = "fao")
}
\author{
Gutama Girja Urago
}
