% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.R
\name{dummify}
\alias{dummify}
\title{Convert a Variables into Dummy}
\usage{
dummify(data, vars, refs = NULL, keep = FALSE)
}
\arguments{
\item{data}{is a data frame containing variables.}

\item{vars}{is a vector of variables to be coded into dummy.}

\item{refs}{is a (named) vector used as reference for dummy coding.}

\item{keep}{is a logical indicating if the original variable is also returned.}
}
\value{
A data frame with newly added dummy variables with labels attribute.
}
\description{
It converts categorical or factor variable into binary variable coded as 0 and 1.
}
\details{
It computes a dummy variable for each unique value of the supplied variable
except for reference, if provided.
}
\examples{

dummify(data = iris, vars = Species, refs = c(Species = "setosa"))

}
\author{
Gutama Girja Urago
}
