% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_f.R
\name{extract_f}
\alias{extract_f}
\title{Extract f stat from anova tests}
\usage{
extract_f(models, .round = 2, .stars = c(0.01, 0.05, 0.1))
}
\arguments{
\item{models}{a named list of linear models (lm)}

\item{.round}{an integer to F stat}

\item{.stars}{a vector of p value cutoffs for significance stars}
}
\value{
A named vector to be inserted in summary tables
}
\description{
Extract f stat from anova tests
}
\examples{
if (FALSE){
extract_f(models = list(
Age = stats::lm(age ~ sex, weights = pw, data = my_data)))
}
}
\author{
Gutama Girja Urago
}
