% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{Reads csv survey file}
\usage{
import_csv(data, excel = NULL, sheet = NULL, format = TRUE)
}
\arguments{
\item{data}{is a path to the csv data frame}

\item{excel}{an excel file containing names and labels of variables}

\item{sheet}{specify sheet name if excel has more than one sheet}

\item{format}{Cleans preceding integers in strings and converts to lower}
}
\value{
a data table
}
\description{
Imports csv micro-data file from local directory. I should be
supplemented with the excel workbook that follows
\code{microData::dictionary()} structure.
}
\details{
The column names of the excel file containing the name and description
of variables should be named "var", "name" and "desc". If the file has
been created using this package, nothing to worry about as it automatically
set these names.
}
\examples{
# import_csv("test.csv")

}
\author{
Gutama Girja Urago
}
