% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Multiple Data Files}
\usage{
import_data(
  path = ".",
  pattern = NULL,
  fun = data.table::fread,
  rm_ext = TRUE,
  ...
)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds
to the working directory, \code{\link[base:getwd]{getwd()}}.
Tilde expansion (see \code{\link[base:path.expand]{path.expand()}}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}

\item{pattern}{an optional \code{\link[base:regex]{regular expression}}.
Only file names which match
the regular expression will be returned.}

\item{fun}{a function used to read data files such as \code{\link[data.table:fread]{data.table::fread}}}

\item{rm_ext}{a logical indicating if the extensions are excluded from the
name of list elements to be returned.}

\item{...}{an optional arguments passed to a function used to read data files.}
}
\value{
A list of data frames
}
\description{
Imports multiple data files from single directory and returns a list.
}
\examples{
if (FALSE) {
import_data(path = "data", pattern = "\\\\.csv$", fun = data.table::fread)}
}
