% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mdt.R
\name{import_mdt}
\alias{import_mdt}
\title{Import Microdata}
\usage{
import_mdt(
  mdt,
  metadata,
  old_name = "old_name",
  new_name = "new_name",
  label = "label",
  all_vars = FALSE
)
}
\arguments{
\item{mdt}{a data frame which is microdata}

\item{metadata}{a data frame that contains old names, new names and
labels of each variables}

\item{old_name}{a column name in the \code{metadata} data frame that contains
old names in the microdata. If there is any \code{NA} values these variables are omitted.}

\item{new_name}{a column name in the \code{metadata} data frame that contains
new names to set to the variables in the \code{mdt}}

\item{label}{a column name in the \code{metadata} data frame that contains
labes to be set to the variables in the \code{mdr}}

\item{all_vars}{a logical. If \code{FALSE (default)} variables with \code{NA} values
in the \code{new_name} column will be omitted, and if \code{TRUE} all variables will be returned.}
}
\value{
A data frame with attributes
}
\description{
Import microdata file, set names and labels.
}
\examples{
if(FALSE){
import_mdt(mdt = microdata,
           metadata = metadata.xlsx,
           old_name = "Names",
           new_name = "New Names",
           label = "Labels")
}
}
