% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_entries.R
\name{latest_entries}
\alias{latest_entries}
\title{Latest Catalog Entries}
\usage{
latest_entries(org = "wb", limit = NULL)
}
\arguments{
\item{org}{a string that represents the name of an organization.}

\item{limit}{an integer. Number of latest catalog entries to return.
\code{Default = 15}}
}
\value{
A data frame
}
\description{
Get latest catalog entries
}
\examples{
if (FALSE){
latest_entries("fao", 25)
}
}
\author{
Gutama Girja Urago
}
