% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdt}
\alias{mdt}
\title{Simulated Microdata}
\format{
A data frame with 100 rows and 5 columns:
\describe{
\item{id}{Respondent iID}
\item{v1}{Age}
\item{v2}{Sex}
\item{v3}{Education}
\item{v4}{Salary}
}
}
\usage{
mdt
}
\description{
A randomly generated microdata to demonstrate how some functions work.
}
\keyword{datasets}
