% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_access_types.R
\name{mdt_access_types}
\alias{mdt_access_types}
\title{List data access types}
\usage{
mdt_access_types(org = "wb")
}
\arguments{
\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}
}
\value{
A data frame
}
\description{
Gets all data access types
}
\examples{
mdt_access_types()

}
