% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_clean.R
\name{mdt_clean}
\alias{mdt_clean}
\title{Cleans Microdata}
\usage{
mdt_clean(
  mdt,
  metadata,
  old_name = "old_name",
  new_name = "new_name",
  label = "label",
  all_vars = FALSE
)
}
\arguments{
\item{mdt}{a dataframe}

\item{metadata}{a dataframe that contains old names, new names and
labels of each variables}

\item{old_name}{a column name in the \code{metadata} data frame that contains
old names in the microdata. If there is any \code{NA} values these variables are omitted.}

\item{new_name}{a column name in the \code{metadata} data frame that contains
new names to set to the variables in the \code{mdt}}

\item{label}{a column name in the \code{metadata} data frame that contains
labes to be set to the variables in the \code{mdr}}

\item{all_vars}{a logical. If \code{FALSE (default)} variables with \code{NA} values
in the \code{new_name} column will be omitted, and if \code{TRUE} all variables will be returned.}
}
\value{
A dataframe with attributes
}
\description{
Cleans microdata file by setting names and labels.
}
\examples{
if(FALSE){
mdt_clean(mdt = microdata,
           metadata = metadata,
           old_name = "Names",
           new_name = "New Names",
           label = "Labels")
}
}
\author{
Gutama Girja Urago
}
