% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_collection.R
\name{mdt_collection}
\alias{mdt_collection}
\title{Get a collection}
\usage{
mdt_collection(id, org = "wb")
}
\arguments{
\item{id}{is an integer repository id.
See object returned by \code{mdt_collections()}.}

\item{org}{is a string of "wb", "fao", "unhcr" or "ilo" organization.}
}
\value{
Prints to console
}
\description{
Gets and prints a description about single collection.
}
\examples{
mdt_collection(26)
}
