% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_country_codes.R
\name{mdt_country_codes}
\alias{mdt_country_codes}
\title{Get a list of all country codes}
\usage{
mdt_country_codes(org = "wb")
}
\arguments{
\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}
}
\value{
A data frame with all country codes
}
\description{
Get a list of all country codes
}
\examples{
mdt_country_codes()
}
