% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_data_files.R
\name{mdt_data_files}
\alias{mdt_data_files}
\title{Get study data file lists}
\usage{
mdt_data_files(idno, org = "wb")
}
\arguments{
\item{idno}{A string that identifies the study. See object returned by \code{mdt_seach()} or \code{mdt_latest()}}

\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}
}
\value{
a data frame
}
\description{
Get study data file lists
}
\examples{
mdt_data_files("ALB_2012_LSMS_v01_M_v01_A_PUF", "wb")
}
