% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_files.R
\name{mdt_files}
\alias{mdt_files}
\title{Study Data Files}
\usage{
mdt_files(id, org = "wb")
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.
See object returned by \code{mdt_seach()} or \code{mdt_latest()}}

\item{org}{a string that represents the name of an organization.}
}
\value{
A data frame
}
\description{
Get list of all study data files
}
\examples{
mdt_files("ALB_2012_LSMS_v01_M_v01_A_PUF", "wb")
}
