% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_latest.R
\name{mdt_latest}
\alias{mdt_latest}
\title{Latest catalog entries}
\usage{
mdt_latest(org = "wb")
}
\arguments{
\item{org}{is a string of "wb", "fao" or "unhcr" organization.}
}
\value{
A data frame
}
\description{
Gets all country codes
}
\examples{
mdt_latest()
}
