% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_search.R
\name{mdt_search}
\alias{mdt_search}
\title{Search microdata catalog}
\usage{
mdt_search(
  keyword = NULL,
  org = "wb",
  from = NULL,
  to = NULL,
  country = NULL,
  inc_iso = FALSE,
  collection = NULL,
  created = NULL,
  dtype = NULL,
  sort_by = NULL,
  sort_order = NULL,
  results = NULL,
  page = NULL,
  detailed = FALSE
)
}
\arguments{
\item{keyword}{A string keyword used to search}

\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}

\item{from}{An integer indicating start year for the data collection.}

\item{to}{An integer indicating end year for the data collection.}

\item{country}{A string. Provide country name or ISO 3 code.
For country names and codes, see \code{mdt_country_codes()}.
For searching multiple countries, you can mix country name and ISO 3 code.
Example \code{c("Afghanistan", "Indonesia")} or \code{c("afg", "ind")} or \code{c("afghanistan", "ind")}}

\item{inc_iso}{A logical. Set the parameter value to \code{TRUE} to
include the ISO3 country codes in the results. \code{Default = FALSE}}

\item{collection}{A string. Filter results by one or more collections.}

\item{created}{A string. Filter results by date of creation.
Use the date format YYYY-MM-DD. Examples, 2020/04/01 returns records created on and after
that date. To specify a date range, use the format 2020/04/01-2020/04/20}

\item{dtype}{A string. Filter results by one or more data access types and takes values of
"open", "direct", "public", "licensed", "enclave", "remote" and "other".
For example, \code{c("open", "direct")}. See \code{mdt_access_types()}}

\item{sort_by}{A string value used to sort the result and takes values of
"rank", "title", "country" and "year" Sort search results.}

\item{sort_order}{A string indicating how to sort the result and takes values of
"asc" or "desc". Ascending = asc, Descending = desc}

\item{results}{An integer indicating number of records per page.
set \code{results = "all"} to get all found results. Default is 15 records per page.}

\item{page}{An integer. Page to return for the search result}

\item{detailed}{A logical. if \code{TRUE} returns a list with all details of
search results, otherwise returns a data frame with only key information.
Default \code{detailed = FALSE}}
}
\value{
A list or data frame based on the \code{detailed} argument.
}
\description{
Search microdata catalog
}
\examples{
mdt_search("lsms")
}
