% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_study_info.R
\name{mdt_study_info}
\alias{mdt_study_info}
\title{Get study info}
\usage{
mdt_study_info(idno, org = "wb")
}
\arguments{
\item{idno}{A string that identifies the study. See object returned by \code{mdt_seach()} or \code{mdt_latest()}}

\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}
}
\value{
a list
}
\description{
Get study info
}
\examples{
mdt_study_info("AFG_2022_BM-UBC_v01_M", "wb")
}
