% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_variable.R
\name{mdt_variable}
\alias{mdt_variable}
\title{Variable Metadata}
\usage{
mdt_variable(id, var_id, org = "wb")
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.}

\item{var_id}{a string specifying variable}

\item{org}{a string specifying organization}
}
\value{
A list
}
\description{
Variable Metadata
}
\examples{
mdt_variable(6161, "v10")
}
