% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_variables.R
\name{mdt_variables}
\alias{mdt_variables}
\title{Variables in the File}
\usage{
mdt_variables(id, file_id, org = "wb")
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.}

\item{file_id}{a string specifying file id}

\item{org}{a string specifying organization}
}
\value{
A data frame
}
\description{
Get list of all variables included in a single file
}
\examples{
mdt_variables(6161, "F1")
}
\author{
Gutama Girja Urago
}
