% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_variables_all.R
\name{mdt_variables_all}
\alias{mdt_variables_all}
\title{Variables in the Study}
\usage{
mdt_variables_all(id, org = "wb")
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.}

\item{org}{a string specifying organization}
}
\value{
A data frame
}
\description{
Get all lists of variables in a study
}
\examples{
mdt_variables_all(6161)
}
\author{
Gutama Girja Urago
}
