% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Metadata of the datasets}
\usage{
metadata(catalog, org = "wb")
}
\arguments{
\item{catalog}{An integer number located in the link after \verb{catalog/} that
identifies the catalog}

\item{org}{Name of an organization. \code{default = 'wb'}, also supports 'fao', 'unchr' and 'ihsn'.}
}
\value{
A data table
}
\description{
Extracts names of available datasets along with their description and dimensions
}
\examples{
metadata(1970)
}
