% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microData-package.R
\docType{package}
\name{microData-package}
\alias{microData}
\alias{microData-package}
\title{microData: Search, Browse, and Import Metadata for Microdata}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An API wrapper package that helps to search, browse, and import metadata for microdata published by international organizations. At the moment, the package supports microdata available on the repository of the WB, FAO, ILO, and IHSN. These organizations use the NADA REST API to publish collected surveys with different types of access. In addition, the package includes some miscellaneous but important functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GutUrago/microData}
  \item Report bugs at \url{https://github.com/GutUrago/microData/issues}
}

}
\author{
\strong{Maintainer}: Gutama Girja Urago \email{girjagutama@gmail.com} (\href{https://orcid.org/0000-0001-5588-2301}{ORCID}) [copyright holder]

}
\keyword{internal}
