% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdt}
\alias{mtdt}
\title{An Example of Metadata}
\format{
A data frame with 100 rows and 5 columns:
\describe{
\item{var_id}{An existing name of variable}
\item{var_name}{Meaningful name of variable}
\item{label}{Labels of variable}
}
}
\usage{
mtdt
}
\description{
An example metadata for a randomly generated microdata to demonstrate how some functions work.
}
\keyword{datasets}
