% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{surveys}
\alias{surveys}
\title{Available surveys}
\usage{
surveys(org = "wb")
}
\arguments{
\item{org}{ame of an organization. \code{default = 'wb'}, also supports 'fao', 'unchr' and 'ihsn'.}
}
\value{
a data table
}
\description{
Extracts the available surveys from the organization. Returns name of the country,
year, catalog and title of each survey.
}
\examples{
surveys("unhcr")
}
\author{
Gutama Girja Urago
}
