% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anova.R
\name{test_anova}
\alias{test_anova}
\title{Extract F-stat from ANOVA Test}
\usage{
test_anova(
  data,
  vars,
  by,
  w = NULL,
  .round = 2,
  .stars = c(0.01, 0.05, 0.1),
  ...
)
}
\arguments{
\item{data}{a data frame containing all variables passed to arguments}

\item{vars}{a string or (named) vector of continuous variables}

\item{by}{a string name of categorical variable}

\item{w}{a string of weights variable name}

\item{.round}{an integer value and digits after decimal}

\item{.stars}{a vector of significance p values}

\item{...}{other arguments to pass to \code{stats::lm()} model}
}
\value{
A named vector of F values with significance stars
}
\description{
For each continuous variable supplied, it extracts F stat (ANOVA) and
returns named vector.
}
\examples{
if(FALSE) {
test_anova(data = iris,
vars = c("Sepal length" = "Sepal.Length",
"Sepal Width" = "Sepal.Width"),
by = "Species", .round = 3)
}
}
\author{
Gutama Girja Urago
}
