% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mean.R
\name{test_mean}
\alias{test_mean}
\title{Extract Mean Difference}
\usage{
test_mean(
  data,
  vars,
  by,
  w = NULL,
  .round = 2,
  .stars = c(0.01, 0.05, 0.1),
  ...
)
}
\arguments{
\item{data}{a data frame containing all variables provided in the arguments}

\item{vars}{a string and (named) vector of numerical variables}

\item{by}{a string name of categorical (factor) variable with two levels}

\item{w}{a string name of weight variable}

\item{.round}{an integer of digits after decimal point for mean differences}

\item{.stars}{a numerical vector corresponding to \verb{***}, \verb{**}, and \code{*}
significance stars, respectively}

\item{...}{arguments passed to \code{weights::wtd.t.test()}}
}
\value{
Returns a named vector
}
\description{
It extracts multiple (weighted) mean difference along with t-test
significance stars using \code{weights} packages.
}
\examples{
if (FALSE) {
test_mean(data = my_data,
vars = c(Age = "age", Income = "income"),
by = "gender",
w = "pw")
}
}
\author{
Gutama Girja Urago
}
