% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dummy.R
\name{to_dummy}
\alias{to_dummy}
\title{Reduce to binary factor}
\usage{
to_dummy(x, look = "yes", yes = "yes", no = "no")
}
\arguments{
\item{x}{a variable}

\item{look}{targeted values}

\item{yes}{values set to target values}

\item{no}{values set to if the test is false}
}
\value{
factor with two levels
}
\description{
Used with grouped data frames to aggregate variables
by testing if the value \code{look} observed at least once
}
\examples{
# to_dummy(x)
}
