% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_workbook.R
\name{to_workbook}
\alias{to_workbook}
\title{Add data frame to workbook}
\usage{
to_workbook(data, wb, sheet)
}
\arguments{
\item{data}{data to be written}

\item{wb}{Workbook}

\item{sheet}{sheetname}
}
\value{
updated workbook
}
\description{
Add data frame to workbook
}
\examples{
# dictionary(3823, "F2") |> to_workbook(wb, "sheet 1")
}
