% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable}
\alias{variable}
\title{Variable Metadata}
\usage{
variable(id, var_id, org = c("wb", "fao", "unhcr", "ihsn", "ilo"))
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.}

\item{var_id}{a string specifying variable}

\item{org}{a string specifying organization}
}
\value{
A list
}
\description{
Get a single variable metadata
}
\examples{
if (FALSE){
variable(6161, "v10")
}
}
\author{
Gutama Girja Urago
}
