% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Variables in the File}
\usage{
variables(id, file_id, org = c("wb", "fao", "unhcr", "ihsn", "ilo"))
}
\arguments{
\item{id}{a study id number. Accepts any of \code{id} (integer) and
\code{idno} (string) values to fetch study information.}

\item{file_id}{a string specifying file id}

\item{org}{a string specifying organization}
}
\value{
A data frame
}
\description{
Get list of all variables included in a single file
}
\examples{
if (FALSE){
variables(6161, "F1")
}
}
\author{
Gutama Girja Urago
}
