% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_t_test.R
\name{wtd_t_test}
\alias{wtd_t_test}
\title{Weighted Students t-Test}
\usage{
wtd_t_test(
  data,
  x,
  y = NULL,
  w = NULL,
  w.y = NULL,
  alternative = c("two.sided", "less", "greater"),
  paired = FALSE,
  mu = 0,
  conf.level = 0.95,
  var.equal = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{an optional matrix or data frame placed as first argument for the
function to be pipe-friendly.}

\item{x}{a name of numeric variable in the data frame or numeric vector in the environment.}

\item{y}{an optional name variable in the data frame or a vector in the environment.
It can be a numeric or factor with two levels. See details.}

\item{w}{an optional (weights) name of numeric variable in the data frame or in the environment.
See details.}

\item{w.y}{an optional weights argument for y.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}". You can
specify just the initial letter.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{mu}{a number indicating the true value of the mean (or difference in
means if you are performing a two sample test).}

\item{conf.level}{onfidence level of the interval.}

\item{var.equal}{a logical variable indicating whether to treat the two
variances as being equal. If TRUE then the pooled variance is used to
estimate the variance otherwise the Welch (or Satterthwaite) approximation
to the degrees of freedom is used.}

\item{na.rm}{a logical indicating wether you want to remove \code{NA} values.}
}
\value{
A list with class "\code{htest}" containing the following components:
\itemize{
\item \code{statistic}: the value of the t-statistic.
\item \code{parameter}: the degrees of freedom for the t-statistic.
\item \code{p.value}: the p-value for the test.
\item \code{conf.int}: a confidence interval for the mean appropriate to the specified alternative hypothesis.
\item \code{estimate}: the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.
\item \code{null.value}: the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.
\item \code{stderr}: the standard error of the mean (difference), used as denominator in the t-statistic formula.
\item \code{alternative}: a character string describing the alternative hypothesis.
\item \code{method}: a character string indicating what type of t-test was performed.
\item \code{data.name}: a character string giving the name(s) of the data.
}
}
\description{
Performs a one-sample, a two-sample and a paired weighted t-test. It is
a slightly modified version of \code{\link[stats:t.test]{t.test}} with added functionality
of supporting weight and pipe.
}
\details{
It shares the same default behavior with \code{\link[stats:t.test]{t.test}}, except
\code{var.equal = TRUE} for this function.

\code{alternative = "greater"} is the alternative that \code{x} has a larger mean than \code{y}.
For the one-sample case: that the mean is positive.

If \code{paired} is \code{TRUE} then both \code{x} and \code{y} must be specified and they must be the same length. Missing
values are silently removed (in pairs if \code{paired} is \code{TRUE}). By default,
if \code{var.equal} is \code{TRUE} then the variance is estimated separately for both
groups and the Welch modification to the degrees of freedom is used.

If the input data are effectively constant (compared to the larger of the
two means) an error is generated.
}
\examples{
library(datasets)
# One sample t-test
wtd_t_test(mtcars, disp, mu = 200)
# Variables can be a vector outside of data frame
wtd_t_test(mtcars, disp, w = rep(c(1, 2), 16), mu = 200)
# Paired t-test
wtd_t_test(mtcars, gear, carb, paired = TRUE)
# Two sample t-test
wtd_t_test(mtcars, drat, wt)
wtd_t_test(mtcars, mpg, as.factor(vs))
wtd_t_test(mtcars, mpg, as.factor(vs), w = qsec)
}
\references{
This function if a modified version of \link[stats:t.test]{t.test}.
}
\author{
Gutama Girja Urago
}
