% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{wtd_var}
\alias{wtd_var}
\alias{variance}
\title{Weighted Variance}
\usage{
variance(x, w = NULL, data = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{a name of numeric variable in the data frame or numeric vector in the environment.}

\item{w}{aan optional (weights) name of numeric variable in the data frame or in the environment.}

\item{data}{an optional matrix or data frame placed as first argument for the
function to be pipe-friendly.}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A numeric vector of length 1.
}
\description{
Calculates weighted variance and standard deviation in a pipe-friendly way.
}
\examples{
variance(height, w = w, data = micro_data)
}
