% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_types.R
\name{access_types}
\alias{access_types}
\title{Data Access Types}
\usage{
access_types(org = "wb")
}
\arguments{
\item{org}{a string that represents the name of an organization.}
}
\value{
A data frame
}
\description{
Get all access types for data
}
\examples{
access_types()

}
