% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.R
\name{dummify}
\alias{dummify}
\title{Dummify variables}
\usage{
dummify(.data, vars, refs = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{vars}{a vector of variables to be used to create dummmies}

\item{refs}{a reference level for variables}
}
\value{
A data frame with newly added dummy variables
}
\description{
Dummifies variables in a data set and add new dummies to a data frame
}
\examples{
if (FALSE){
dummify(df, c("sex", "race"), c("m", "b"))
}
}
\author{
Gutama Girja Urago
}
