% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.R
\name{dummify}
\alias{dummify}
\title{Convert a Variables into Dummy}
\usage{
dummify(data, ..., var_lev = TRUE, keep = TRUE)
}
\arguments{
\item{data}{is a matrix or data frame containing variables.}

\item{...}{a names of variables to be coded into dummy.}

\item{var_lev}{is a logical indicating if new variables are named as
'variable_level' style or just only levels.}

\item{keep}{is a logical indicating if the original variable is also returned.}
}
\value{
A data frame with newly added dummy variables with labels attribute.
}
\description{
It converts categorical or factor variable into binary variable coded as 0 and 1.
}
\details{
It computes a dummy variable for each unique value of the supplied variable
except for reference, if provided.
}
\examples{

dummify(data = iris, Species)
# Work on how to include vectors without quotation
}
\author{
Gutama Girja Urago
}
