% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_f.R
\name{extract_f}
\alias{extract_f}
\title{Extract F stat from lm}
\usage{
extract_f(x, by, data, .round = 2, .stars = c(0.01, 0.05, 0.1), ...)
}
\arguments{
\item{x}{a vector of continuous variable}

\item{by}{a character name of categorical variable}

\item{data}{a data frame}

\item{.round}{an integer value and digits after decimal}

\item{.stars}{a vector of significance p values}

\item{...}{other arguments to pass to \code{stats::lm} model}
}
\value{
A named vector of F values with significance stars
}
\description{
For each continuous variable supplied, it extracts F stat (ANOVA) and
returns named vector.
}
\examples{
if(FALSE) {
extract_f(x = c("Sepal length" = "Sepal.Length"),
by = "Species", data = iris, .round = 3)
}
}
\author{
Gutama Girja Urago
}
