% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Multiple Data Files}
\usage{
import_data(
  path = ".",
  pattern = NULL,
  fun = data.table::fread,
  rm_ext = TRUE,
  ...
)
}
\arguments{
\item{path}{A character vector specifying the directory path.
Defaults to the working directory, \code{\link[base:getwd]{getwd()}}.
Missing or invalid paths will trigger an error.}

\item{pattern}{An optional \code{\link[base:regex]{regular expression}}
to filter file names. Only files matching the pattern will be processed.}

\item{fun}{A function used to read data files,
such as \code{\link[data.table:fread]{data.table::fread}} or \code{\link[utils:read.csv]{read.csv}}.}

\item{rm_ext}{A logical value indicating whether file extensions should
be removed from the names of the returned list elements. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to the data reading function specified in \code{fun}.}
}
\value{
A named list of data frames, where names correspond to file names (optionally without extensions).
}
\description{
Imports multiple data files from a specified directory and returns a
named list of data frames.
}
\examples{
if (FALSE) {
# Example with data.table::fread
import_data(path = "data", pattern = "\\\\.csv$", fun = data.table::fread)

# Example with base R read.csv
import_data(path = "data", pattern = "\\\\.csv$", fun = read.csv, rm_ext = FALSE)
}
}
