% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_collection.R
\name{mdt_collection}
\alias{mdt_collection}
\title{Get a collection}
\usage{
mdt_collection(repo_id, org = "wb", display = TRUE, html = FALSE)
}
\arguments{
\item{repo_id}{A repository id.
See object returned by \code{mdt_collections()}.}

\item{org}{is a string of "wb", "fao", "unhcr" or "ilo" organization.}

\item{display}{A logical. If \code{TRUE} prints html. \code{Default = TRUE}}

\item{html}{A logical, If \code{TRUE} returns html file. \code{Default = FALSE}}
}
\value{
Prints to console
}
\description{
Gets and prints a description about a specific collection.
}
\examples{
mdt_collection("fao")
}
