% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_collection.R
\name{mdt_collection}
\alias{mdt_collection}
\title{Get A Collection}
\usage{
mdt_collection(repo_id, org = "wb")
}
\arguments{
\item{repo_id}{a repository id.
See object returned by \code{mdt_collections()}.}

\item{org}{a character string specifying the name of an organization
that will return the request.}
}
\value{
Prints to html viewer
}
\description{
Get description about a specific collection.
}
\examples{
mdt_collection("fao")
}
\author{
Gutama Girja Urago
}
