% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_country_codes.R
\name{mdt_country_codes}
\alias{mdt_country_codes}
\title{List of all country codes}
\usage{
mdt_country_codes(org = "wb")
}
\arguments{
\item{org}{is a string of "wb", "fao" or "unhcr" organization.}
}
\value{
A data frame
}
\description{
List of all country codes
}
\examples{
mdt_country_codes()
}
