% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_latest.R
\name{mdt_latest}
\alias{mdt_latest}
\title{Latest catalog entries}
\usage{
mdt_latest(org = "wb", limit = NULL)
}
\arguments{
\item{org}{A string that represents the name of an organization.
At the moment, it supports "wb", "fao", "unhcr", "ihsn" and "ilo" organizations.}

\item{limit}{An integer. Number of latest catalog entries to return.
\code{Default = 15}}
}
\value{
A data frame
}
\description{
Get latest catalog entries
}
\examples{
mdt_latest("fao", 25)
}
