% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_latest.R
\name{mdt_latest}
\alias{mdt_latest}
\title{Latest Catalog Entries}
\usage{
mdt_latest(org = "wb", limit = NULL)
}
\arguments{
\item{org}{a string that represents the name of an organization.}

\item{limit}{an integer. Number of latest catalog entries to return.
\code{Default = 15}}
}
\value{
A data frame
}
\description{
Get latest catalog entries
}
\examples{
mdt_latest("fao", 25)
}
