% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_search.R
\name{mdt_search}
\alias{mdt_search}
\title{Search micro data catalog}
\usage{
mdt_search(
  org = "wb",
  from = NULL,
  to = NULL,
  country = NULL,
  collection = NULL,
  dtype = NULL,
  sort_by = NULL,
  sort_order = NULL,
  ps = NULL,
  page = NULL,
  detailed = FALSE
)
}
\arguments{
\item{org}{is a string of "wb", "fao" or "unhcr" organization.}

\item{from}{is an integer start year for the data collection.}

\item{to}{is an integer end year for the data collection.}

\item{country}{is a string. Provide country name or ISO 3 code.
For country names and codes, see \code{mdt_country_codes()}.
For searching multiple countries, provide names separated by pipe(|).
Example Afghanistan|Indonesia or afg|ind or afghanistan|ind}

\item{collection}{Filter results by one or more collections.
For filtering multiples, use format collection1, collection2}

\item{dtype}{a string. Data access types and takes values of
"open""direct""public""licensed""enclave""remote""other"
Filter by data access types. For multiple values,
use comma e.g. open, direct, remote}

\item{sort_by}{a string value used to sort the result and takes values of
"rank""title""nation""year" Sort search results.
Valid options are rank,title,year,nation}

\item{sort_order}{a string indicated how to sort the result and takes values of
"asc" "desc" Set results sort order. Ascending = asc, Descending= desc}

\item{ps}{an integer, Number of results per page.
Default is 15 records per page}

\item{page}{an integer. Page to return for the search result}

\item{detailed}{a logical. if \code{TRUE} returns a list with all
search results, otherwise returns a data frame. Default \code{detailed = FALSE}}
}
\value{
A list or data frame based on the \code{detailed} argument.
}
\description{
A short description...
}
\examples{
mdt_search()
}
