% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_t_test.R
\name{wtd_t_test}
\alias{wtd_t_test}
\alias{wtd_t_test.default}
\alias{wtd_t_test.formula}
\title{Weighted Students t-Test}
\usage{
wtd_t_test(x, ...)

\method{wtd_t_test}{default}(
  x,
  y = NULL,
  w = NULL,
  w.y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = TRUE,
  conf.level = 0.95,
  ...
)

\method{wtd_t_test}{formula}(formula, w = NULL, data, subset, na.action = na.pass, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{Additional arguments passed to or from other methods.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{w}{an optional numeric vector of weights for \code{x}. If \code{NULL}, equal weights are assumed.}

\item{w.y}{an optional numeric vector of weights for \code{y}. If \code{NULL}, equal weights are assumed.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}". You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test. Default is \code{FALSE}.}

\item{var.equal}{a logical variable indicating whether to treat the two variances
as being equal. If \code{TRUE} (default) then the pooled variance is used to estimate
the variance otherwise the Welch (or Satterthwaite) approximation to the
degrees of freedom is used.}

\item{conf.level}{confidence level of the interval. Default is 0.95.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs} is a numeric variable
giving the data values and \code{rhs} either \code{1} for a one-sample or paired test or a
factor with two levels giving the corresponding groups. If \code{lhs} is of class \link[stats:Pair]{"Pair"}
and \code{rhs} is \code{1}, a paired test is done, see Examples.}

\item{data}{an optional matrix or data frame (or similar: see \link[stats:model.frame]{model.frame})
containing the variables in the formula formula. By default the variables are
taken from \code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain \link[base:NA]{NA}s.}
}
\value{
A list with class "\code{htest}" containing the following components:
\itemize{
\item \code{statistic}: the value of the t-statistic.
\item \code{parameter}: the degrees of freedom for the t-statistic.
\item \code{p.value}: the p-value for the test.
\item \code{conf.int}: a confidence interval for the mean appropriate to the specified alternative hypothesis.
\item \code{estimate}: the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.
\item \code{null.value}: the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.
\item \code{stderr}: the standard error of the mean (difference), used as denominator in the t-statistic formula.
\item \code{alternative}: a character string describing the alternative hypothesis.
\item \code{method}: a character string indicating what type of t-test was performed.
\item \code{data.name}: a character string giving the name(s) of the data.
}
}
\description{
Performs a one-sample, a two-sample and a paired weighted t-test. It is
a slightly modified version of \code{\link[stats:t.test]{t.test}} with added functionality
of supporting weight.
}
\details{
It shares the same default behavior with \code{\link[stats:t.test]{t.test}}, except
\code{var.equal = TRUE} for this function.

\code{alternative = "greater"} is the alternative that \code{x} has a larger mean than \code{y}.
For the one-sample case: that the mean is positive.

If \code{paired} is \code{TRUE} then both \code{x} and \code{y} must be specified and they must be the same length. Missing
values are silently removed (in pairs if \code{paired} is \code{TRUE}). By default,
if \code{var.equal} is \code{TRUE} then the variance is estimated separately for both
groups and the Welch modification to the degrees of freedom is used.

If the input data are effectively constant (compared to the larger of the
two means) an error is generated.
}
\examples{
# Example 1: One-sample t-test
x <- rnorm(10)
w <- runif(10)
wtd_t_test(x, w = w)

# Example 2: Two-sample t-test with equal variances
x <- rnorm(10)
y <- rnorm(10)
wtd_t_test(x, y)

# Example 3: Paired t-test
x <- rnorm(10)
y <- x + rnorm(10)
wtd_t_test(x, y, paired = TRUE)
}
\references{
This function if a modified version of \link[stats:t.test]{t.test}.
}
\author{
Gutama Girja Urago
}
