% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_var.R
\name{wtd_var}
\alias{wtd_var}
\alias{wtd_sd}
\title{Weighted Variance}
\usage{
wtd_var(x, w, na.rm = TRUE)

wtd_sd(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{an optional numeric vector of weights}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A numeric vector of length 1.
}
\description{
Calculates weighted variance and standard deviation.
}
\examples{
x <- rnorm(100)
w <- rep(c(1,2), 50)
wtd_var(x, w)
}
