% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_var.R
\name{wtd_var}
\alias{wtd_var}
\alias{wtd_sd}
\title{Weighted Variance}
\usage{
wtd_var(data, x, w = NULL, na.rm = TRUE)

wtd_sd(data, x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{an optional matrix or data frame placed as first argument for the
function to be pipe-friendly.}

\item{x}{a name of numeric variable in the data frame or numeric vector in the environment.}

\item{w}{aan optional (weights) name of numeric variable in the data frame or in the environment.}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A numeric vector of length 1.
}
\description{
Calculates weighted variance and standard deviation in a pipe-friendly way.
}
\examples{
x <- rnorm(100)
w <- rep(c(1,2), 50)
wtd_var(x = x, w = w)
}
